+++ client/client.rs
@@ -13,7 +13,7 @@ use pueue_lib::network::secret::read_shared_secret;
 use pueue_lib::settings::Settings;
 use pueue_lib::state::PUEUE_DEFAULT_GROUP;
 
-use crate::cli::{CliArguments, SubCommand};
+use crate::cli::{CliArguments, GroupCommand, SubCommand};
 use crate::commands::edit::edit;
 use crate::commands::get_state;
 use crate::commands::local_follow::local_follow;
@@ -468,22 +468,18 @@ impl Client {
                 };
                 Ok(Message::Send(message))
             }
-            SubCommand::Group {
-                add,
-                parallel,
-                remove,
-            } => {
-                if let Some(group) = add {
+            SubCommand::Group { cmd } => match cmd {
+                Some(GroupCommand::Add { name, parallel }) => {
                     Ok(Message::Group(GroupMessage::Add {
-                        name: group.to_owned(),
+                        name: name.to_owned(),
                         parallel_tasks: parallel.to_owned(),
                     }))
-                } else if let Some(group) = remove {
-                    Ok(Message::Group(GroupMessage::Remove(group.to_owned())))
-                } else {
-                    Ok(Message::Group(GroupMessage::List))
                 }
-            }
+                Some(GroupCommand::Remove { name }) => {
+                    Ok(Message::Group(GroupMessage::Remove(name.to_owned())))
+                }
+                None => Ok(Message::Group(GroupMessage::List)),
+            },
             SubCommand::Status { .. } => Ok(Message::Status),
             SubCommand::Log {
                 task_ids,